requirements.txt
praw>=7.7.0
tweepy>=4.14.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
textblob>=0.17.1
schedule>=1.2.0
python-dateutil>=2.8.0